replace manual definition of bin product of two Cartesian categories
just use the total category of weakening definition, so we don't have
two different definitionally different definitions of bin product of
Cartesian categories, that differ only on the proofs of the Cartesian
structure. Anyways, the underlying displayed categories are such that
the total category of the weakening is definitionally the binary product
although they are each defined manually. There isn't an easy way to
define weakening from the bin product so let's define the bin product
from weakening