I think I finally got NormalForm τ τ by doing a stronger induction
I was really stuck on how to justify that forall τ,
there is a NormalForm τ τ

The intuitive "it's just a big ast of projections" seemed hard to
formalize, but I think this lexicographic induction works?
Not sure, still confused a bit by what just happened.

But the equational properties we need now seem like a pain.
Maybe that's self-inflicted by adding isSet to the neutral and normal
forms.

We'll see